create the Texture2DFactory extension
because:
- 2d textures are a really common need
- it doesn't make sense for the gpu atlas to own this behavior when
  other systems will need to create textures
